+++ cmd/launcher/launcher.go
@@ -3,19 +3,16 @@ package main
 import (
 	"bytes"
 	"context"
-	"crypto/sha256"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/json"
 	"flag"
 	"fmt"
 	"io/ioutil"
-	"net"
 	"net/http"
 	"os"
 	"os/signal"
 	"path/filepath"
-	"strings"
 	"syscall"
 	"text/tabwriter"
 	"time"
@@ -38,8 +35,6 @@ import (
 	"github.com/kolide/osquery-go/plugin/distributed"
 	osquery_logger "github.com/kolide/osquery-go/plugin/logger"
 	"github.com/pkg/errors"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials"
 )
 
 var (
@@ -306,7 +301,7 @@ func main() {
 		}
 	}
 
-	conn, err := dialGRPC(opts.kolideServerURL, opts.insecureTLS, opts.insecureGRPC, opts.certPins, rootPool, logger)
+	conn, err := service.DialGRPC(opts.kolideServerURL, opts.insecureTLS, opts.insecureGRPC, opts.certPins, rootPool, logger)
 	if err != nil {
 		logger.Fatal("err", errors.Wrap(err, "dialing grpc server"))
 	}
@@ -433,111 +428,4 @@ func launcherFinalizer(logger log.Logger, shutdownOsquery func() error) func() e
 		}
 		return nil
 	}
-}
-
-// tempErr is a wrapper for errors that are "temporary" and should be retried
-// for gRPC.
-type tempErr struct {
-	error
-}
-
-func (t tempErr) Temporary() bool {
-	return true
-}
-
-// tlsCreds overwrites the ClientHandshake method for specific error handling.
-type tlsCreds struct {
-	credentials.TransportCredentials
-}
-
-// ClientHandshake wraps the normal gRPC ClientHandshake, but treats a
-// certificate with the wrong name as a temporary rather than permanent error.
-// This is important for reconnecting to the gRPC server after, for example,
-// the certificate being MitMed by a captive portal (without this, gRPC calls
-// will error and never attempt to reconnect).
-// See https://github.com/grpc/grpc-go/issues/1571.
-func (t *tlsCreds) ClientHandshake(ctx context.Context, s string, c net.Conn) (net.Conn, credentials.AuthInfo, error) {
-	conn, info, err := t.TransportCredentials.ClientHandshake(ctx, s, c)
-	if err != nil && strings.Contains(err.Error(), "x509: certificate is valid for ") {
-		err = &tempErr{err}
-	}
-
-	return conn, info, err
-}
-
-// dialGRPC creates a grpc client connection.
-func dialGRPC(
-	serverURL string,
-	insecureTLS bool,
-	insecureGRPC bool,
-	certPins [][]byte,
-	rootPool *x509.CertPool,
-	logger log.Logger,
-	opts ...grpc.DialOption, // Used for overrides in testing
-) (*grpc.ClientConn, error) {
-	level.Info(logger).Log(
-		"msg", "dialing grpc server",
-		"server", serverURL,
-		"tls_secure", insecureTLS == false,
-		"grpc_secure", insecureGRPC == false,
-		"cert_pinning", len(certPins) > 0,
-	)
-	grpcOpts := []grpc.DialOption{
-		grpc.WithTimeout(time.Second),
-	}
-	if insecureGRPC {
-		grpcOpts = append(grpcOpts, grpc.WithInsecure())
-	} else {
-		host, _, err := net.SplitHostPort(serverURL)
-		if err != nil {
-			return nil, errors.Wrapf(err, "split grpc server host and port: %s", serverURL)
-		}
-
-		creds := &tlsCreds{credentials.NewTLS(makeTLSConfig(host, insecureTLS, certPins, rootPool, logger))}
-		grpcOpts = append(grpcOpts, grpc.WithTransportCredentials(creds))
-	}
-
-	grpcOpts = append(grpcOpts, opts...)
-
-	conn, err := grpc.Dial(serverURL, grpcOpts...)
-	return conn, err
-}
-
-func makeTLSConfig(host string, insecureTLS bool, certPins [][]byte, rootPool *x509.CertPool, logger log.Logger) *tls.Config {
-	conf := &tls.Config{
-		ServerName:         host,
-		InsecureSkipVerify: insecureTLS,
-		RootCAs:            rootPool,
-	}
-
-	if len(certPins) > 0 {
-		conf.VerifyPeerCertificate = func(rawCerts [][]byte, verifiedChains [][]*x509.Certificate) error {
-			for _, chain := range verifiedChains {
-				for _, cert := range chain {
-					// Compare SHA256 hash of
-					// SubjectPublicKeyInfo with each of
-					// the pinned hashes.
-					hash := sha256.Sum256(cert.RawSubjectPublicKeyInfo)
-					for _, pin := range certPins {
-						if bytes.Equal(pin, hash[:]) {
-							// Cert matches pin.
-							return nil
-						}
-					}
-				}
-			}
-
-			// Normally we wouldn't log and return an error, but
-			// gRPC does not seem to expose the error in a way that
-			// we can get at it later. At least this provides some
-			// feedback to the user about what is going wrong.
-			level.Info(logger).Log(
-				"msg", "no match found with pinned certificates",
-				"err", "certificate pin validationf failed",
-			)
-			return errors.New("no match found with pinned cert")
-		}
-	}
-
-	return conf
 }
